ModelApi/requirements.txt
azure-functions
requests
joblib
pandas
scikit-learn
